feat: add home categories and images to categories

+++ src/Mymarket.Application/Features/Categories/Commands/Add/AddCategoryCommand.cs
@@ -5,6 +5,7 @@ using Mymarket.Application.Interfaces;
 using Mymarket.Domain.Enums;
 using Mymarket.Domain.Entities;
 using Microsoft.AspNetCore.Http;
+using Mymarket.Application.Resources;
 
 namespace Mymarket.Application.Features.Categories.Commands.Add;
 
@@ -13,7 +14,8 @@ public record AddCategoryCommand(
     string Name,
     string? NameEn,
     string? NameRu,
-    bool? BrandRequired,
+    bool BrandRequired,
+    bool BrandVisible,
     IFormFile? Logo,
     CategoryPostType CategoryPostType
 ) : IRequest<CategoryDto>;
@@ -26,14 +28,14 @@ public class AddCategoryCommandHandler(
         AddCategoryCommand request, CancellationToken cancellationToken)
     {
         await using var transaction = await context.BeginTransactionAsync(cancellationToken);
-        var uploadedImage = await imageService.UploadAsync(request.Logo, cancellationToken);
+
+        ImageEntity? uploadedImage = null;
 
         try
         {
-            if (request.ParentId.HasValue)
+            if (request.Logo is not null)
             {
-                var parentExists = await context.Categories
-                    .AnyAsync(c => c.Id == request.ParentId.Value, cancellationToken);
+                uploadedImage = await imageService.UploadAsync(request.Logo, cancellationToken);
             }
 
             var category = new CategoryEntity
@@ -43,12 +45,15 @@ public class AddCategoryCommandHandler(
                 NameEn = request.NameEn,
                 NameRu = request.NameRu,
                 BrandRequired = request.BrandRequired,
+                BrandVisible = request.BrandVisible,
                 CategoryPostType = request.CategoryPostType,
-                Logo = new ImageEntity 
-                { 
-                    UniqueId = uploadedImage.UniqueId, 
-                    Url = uploadedImage.Url,
-                }
+                Logo = uploadedImage is null 
+                    ? null 
+                    : new ImageEntity
+                        {
+                            UniqueId = uploadedImage.UniqueId,
+                            Url = uploadedImage.Url,
+                        }
             };
 
             await context.Categories.AddAsync(category, cancellationToken);
@@ -64,11 +69,14 @@ public class AddCategoryCommandHandler(
                 NameEn = category.NameEn,
                 NameRu = category.NameRu,
                 BrandRequired = request.BrandRequired,
+                BrandVisible = request.BrandVisible,
                 CategoryPostType = category.CategoryPostType,
-                LogoUrl = context.Images
-                    .Where(i => i.Id == uploadedImage.Id)
-                    .Select(i => i.Url)
-                    .FirstOrDefault() ?? string.Empty
+                LogoUrl = uploadedImage is null 
+                    ? null
+                    : context.Images
+                        .Where(i => i.Id == category.LogoId)
+                        .Select(i => i.Url)
+                        .FirstOrDefault() ?? string.Empty
             };
 
             return dto;
@@ -76,7 +84,12 @@ public class AddCategoryCommandHandler(
         catch
         {
             await transaction.RollbackAsync(cancellationToken);
-            await imageService.DeleteAsync(uploadedImage, cancellationToken);
+
+            if (uploadedImage is not null)
+            {
+                await imageService.DeleteAsync(uploadedImage, cancellationToken);
+            }
+
             throw;
         }
     }
